refactor(gateway): restructure API and move core modules to align with clean architecture
Reorganized the project structure by moving feature-specific controllers, routes, types, and services into new domain-based and API directories. Core utilities and middleware were relocated to shared and application/api folders, and new placeholder DTO directories were added.
