+++ gateway/src/application/api/analyzed-transaction/analyzed-transaction.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { getTransactions } from "./dashboard.controller.ts";
+import { getTransactions } from "./analyzed-transaction.controller.ts";
 
 const dashboardRoutes = Router();
 dashboardRoutes.get("/transactions", getTransactions);
