+++ gateway/src/application/api/new-transaction/new-transaction.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { processNewTransaction } from "./transactions.controller.ts";
+import { processNewTransaction } from "./new-transaction.controller.ts";
 
 const transactionsRoutes = Router();
 transactionsRoutes.post("/process", processNewTransaction);
