+++ gateway/src/core/analyzed-transaction/schemas/analyzed-transaction.types.ts
