+++ gateway/src/core/new-transaction/schemas/transactions.types.ts
